refactor: move existing tests and their fixtures into a shared directory
