+++ .vscode-test.mjs
@@ -5,11 +5,13 @@
 
 import { defineConfig } from '@vscode/test-cli';
 
+const workspaceFolder = "./src/test/integration/basics/workspace"
+
 const config = defineConfig({
   version: process.env['VSCODE_VERSION'] ?? 'stable',
-  workspaceFolder: process.env['VSCODE_WORKSPACE_FOLDER'] ?? './test/fixtures',
+  workspaceFolder: process.env['VSCODE_WORKSPACE_FOLDER'] ?? workspaceFolder,
   launchArgs: ['--disable-extensions', '--disable-workspace-trust'],
-  files: 'out/test/**/*.test.js',
+  files: 'out/test/integration/basics/*.test.js',
   mocha: {
     ui: 'tdd',
     color: true,
