+++ src/test/integration/basics/hover.test.ts
@@ -5,11 +5,11 @@
 
 import * as vscode from 'vscode';
 import * as assert from 'assert';
-import { activateExtension, getDocUri, open, testSymbols } from '../helper';
+import { activateExtension, getDocUri, open, testHover } from '../../helper';
 
-suite('symbols', () => {
-  suite('basic language symbols', function suite() {
-    const docUri = getDocUri('sample.tf');
+suite('hover', () => {
+  suite('core schema', function suite() {
+    const docUri = getDocUri('main.tf');
 
     this.beforeAll(async () => {
       await open(docUri);
@@ -25,8 +25,15 @@ suite('symbols', () => {
       assert.equal(doc.languageId, 'terraform', 'document language should be `terraform`');
     });
 
-    test('returns symbols', async () => {
-      await testSymbols(docUri, ['provider "vault"', 'resource "vault_auth_backend" "b"', 'module "local"']);
+    test('returns docs for terraform block', async () => {
+      await testHover(docUri, new vscode.Position(0, 1), [
+        new vscode.Hover(
+          new vscode.MarkdownString(
+            '**terraform** _Block_\n\nTerraform block used to configure some high-level behaviors of Terraform',
+          ),
+          new vscode.Range(new vscode.Position(14, 12), new vscode.Position(14, 20)),
+        ),
+      ]);
     });
   });
 });
