+++ src/test/integration/basics/workspace/.vscode/settings.json
@@ -1,0 +1,6 @@
+{
+    "terraform.languageServer.enable": true,
+    "editor.codeActionsOnSave": {
+        "source.formatAll.terraform": true
+    }
+}
