+++ src/test/integration/basics/workspace/actions.tf
@@ -1,0 +1,3 @@
+terraform {
+  # code that needs to be formatted
+                }
