+++ src/test/integration/basics/workspace/ai/variables.tf
@@ -1,0 +1,3 @@
+variable "agi" {
+  default = false
+}
