+++ src/test/integration/basics/workspace/compute/outputs.tf
@@ -1,0 +1,3 @@
+output "ip" {
+  value = google_compute_instance.vm_instance.network_interface[0].network_ip
+}
