+++ src/test/integration/basics/workspace/main.tf
@@ -1,0 +1,24 @@
+terraform {
+  required_providers {
+    google = {
+      source  = "hashicorp/google"
+      version = "~> 4.0"
+    }
+  }
+}
+
+provider "google" {
+  credentials = file(var.credentials_file)
+
+  project = var.project
+  region  = var.region
+  zone    = var.zone
+}
+
+module "compute" {
+  source = "./compute"
+
+  instance_name = "terraform-machine"
+
+}
+
