+++ src/test/integration/basics/workspace/sample.tf
@@ -1,0 +1,9 @@
+provider "vault" {
+}
+
+resource "vault_auth_backend" "b" {
+}
+
+module "local" {
+  source = "./modules"
+}
