+++ src/test/integration/basics/workspace/terraform.tfvars
@@ -1,0 +1,1 @@
+zone = "us-central1-c"
