+++ src/test/integration/basics/workspace/variables.tf
@@ -1,0 +1,15 @@
+variable "project" {
+  type = string
+}
+
+variable "credentials_file" {
+  type = string
+}
+
+variable "region" {
+  default = "us-central1"
+}
+
+variable "zone" {
+  default = "us-central1-c"
+}
